+++ src/cqfantasy/house/__init__.py
@@ -1,3 +1,5 @@
 from .Body import Body
 from .Roof import Roof
 from .House import House
+from .StuccoBrickBody import StuccoBrickBody
+from .TudorSplitBody import TudorSplitBody